+++ klause/src/commonMain/kotlin/com/eignex/klause/factor/scheduling/internals/CumulativeEff.kt
@@ -4,9 +4,9 @@ package com.eignex.klause.factor.scheduling.internals
  *  `null` at the call site when any var-arg is still open — propagation defers in that case. */
 class CumulativeEff(
     /** Fixed durations, one per task. */
-    val dur: IntArray,
+    val dur: LongArray,
     /** Fixed resource requirements, one per task. */
-    val res: IntArray,
+    val res: LongArray,
     /** Capacity ceiling. */
-    val cap: Int,
+    val cap: Long,
 )
